UI, calendar and home filter fixes
UI images have been changed. Scroll view in the home controller cycles
through all the images in the image array. Also the scroll view is like
a pages as well. Added code to fix the issue of the home controller not
loading the home feed. Also fixed the calendar to add the date of an
event. It also checks for duplicates before adding an event and checks
for calendar permissions.